Set the prefix for building with make
Not really necessary currently, but it is nice to set the build prefix
so that any includes or libraries are picked up from our build prefix
and not somewhere else on the system. As openlibm doesn't really have
any build requirements currently, this doesn't affect anything. However
it still nice to have this set correctly here if that should change in
the future. Also makes for a better template when adding other recipes
configured this way.